refactor: moved the files into folders & rename
