+++ .eslintrc.js
@@ -1,7 +1,7 @@
 module.exports = {
   parser: '@typescript-eslint/parser',
   parserOptions: {
-    project: ['./tsconfig.json', './tsconfig.test.json'],
+    project: ['./tsconfig.json'],
     tsconfigRootDir : __dirname, 
     sourceType: 'module',
   },
