+++ src/constants/store.ts
@@ -1,4 +1,4 @@
-import { Store } from './types/types-game';
+import { Store } from '../types/types-game';
 
 export const store: Store = {
   players: [],
