+++ src/utils/game-utils.ts
